# swagger.json
{"openapi":"3.0.3","info":{"title":"ServiceHub API","version":"1.0.0","description":"REST API documentation for the ServiceHub Multi-Vendor Home Services Marketplace. All protected endpoints require a Bearer token in the Authorization header.","contact":{"name":"ServiceHub Engineering","email":"engineering@servicehub.app"},"license":{"name":"ISC"}},"servers":[{"url":"http://localhost:5000/api/v1","description":"Local Development Server"},{"url":"https://api.servicehub.app/api/v1","description":"Production Server"}],"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"Enter your JWT access token"}},"schemas":{},"responses":{"UnauthorizedError":{"description":"Access token is missing or invalid","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Unauthorized"}}}}}},"ForbiddenError":{"description":"Authenticated but not authorized to perform this action","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Forbidden"}}}}}},"NotFoundError":{"description":"The requested resource was not found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Resource not found"}}}}}},"ValidationError":{"description":"Request body failed validation","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Validation failed"},"errors":{"type":"array","items":{"type":"object","properties":{"field":"@{type=string}","message":"@{type=string}"}}}}}}}}}},"security":[{"BearerAuth":[]}],"tags":[{"name":"Health","description":"API health and status"},{"name":"Auth","description":"Authentication — register, login, refresh, logout"},{"name":"Users","description":"Customer profile management"},{"name":"Providers","description":"Provider profile, availability, and services"},{"name":"Admin","description":"Admin-only operations"},{"name":"Categories","description":"Service categories"},{"name":"Services","description":"Service catalog"},{"name":"Bookings","description":"Booking lifecycle management"},{"name":"Payments","description":"Payments and wallet"},{"name":"Reviews","description":"Reviews and ratings"},{"name":"Disputes","description":"Dispute management"},{"name":"Notifications","description":"User notifications"},{"name":"Coupons","description":"Coupon management"}],"paths":{"/wallets/me":{"get":{"summary":"Get current user\u0027s wallet","description":"Retrieves the current balance and recent transactions for the authenticated user\u0027s wallet.","tags":["Payments"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Wallet fetched successfully"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/users/addresses":{"get":{"summary":"GET /users/addresses","tags":["users"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"post":{"summary":"POST /users/addresses","tags":["users"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/users/addresses/{addressId}":{"put":{"summary":"PUT /users/addresses/{addressId}","tags":["users"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"delete":{"summary":"DELETE /users/addresses/{addressId}","tags":["users"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/users/addresses/{addressId}/default":{"patch":{"summary":"PATCH /users/addresses/{addressId}/default","tags":["users"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/reviews":{"get":{"summary":"GET /reviews","tags":["reviews"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"post":{"summary":"POST /reviews","tags":["reviews"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/reviews/{reviewId}/reply":{"post":{"summary":"POST /reviews/{reviewId}/reply","tags":["reviews"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/providers/me":{"get":{"summary":"GET /providers/me","tags":["providers"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"put":{"summary":"PUT /providers/me","tags":["providers"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/notifications":{"get":{"summary":"GET /notifications","tags":["notifications"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/notifications/read-all":{"patch":{"summary":"PATCH /notifications/read-all","tags":["notifications"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/notifications/{notificationId}/read":{"patch":{"summary":"PATCH /notifications/{notificationId}/read","tags":["notifications"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/health":{"get":{"summary":"API Health Check","description":"Returns the operational status of the API and its database connection. This endpoint is public (no authentication required) and is intended for use by uptime monitors and load balancers.\n","tags":["Health"],"security":[],"responses":{"200":{"description":"API is healthy and database is reachable","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"ServiceHub API is running"},"data":{"type":"object","properties":"@{status=; database=; environment=; timestamp=; uptime=}"}}}}}},"503":{"description":"API is up but database is unreachable"}}}},"/disputes":{"get":{"summary":"GET /disputes","tags":["disputes"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"post":{"summary":"POST /disputes","tags":["disputes"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/disputes/{disputeId}":{"get":{"summary":"GET /disputes/{disputeId}","tags":["disputes"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/disputes/{disputeId}/messages":{"post":{"summary":"POST /disputes/{disputeId}/messages","tags":["disputes"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/catalog/categories":{"get":{"summary":"GET /catalog/categories","tags":["catalog"],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/catalog/services":{"get":{"summary":"GET /catalog/services","tags":["catalog"],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/catalog/services/{serviceId}":{"get":{"summary":"GET /catalog/services/{serviceId}","tags":["catalog"],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/bookings":{"post":{"summary":"POST /bookings","tags":["bookings"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}},"get":{"summary":"GET /bookings","tags":["bookings"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/bookings/{bookingId}":{"get":{"summary":"GET /bookings/{bookingId}","tags":["bookings"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/bookings/{bookingId}/status":{"patch":{"summary":"PATCH /bookings/{bookingId}/status","tags":["bookings"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/auth/register":{"post":{"summary":"POST /auth/register","tags":["auth"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/auth/login":{"post":{"summary":"POST /auth/login","tags":["auth"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/auth/me":{"get":{"summary":"GET /auth/me","tags":["auth"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/admin/documents/{docId}/status":{"patch":{"summary":"PATCH /admin/documents/{docId}/status","tags":["admin"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/admin/disputes/{disputeId}/resolve":{"patch":{"summary":"PATCH /admin/disputes/{disputeId}/resolve","tags":["admin"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful operation"},"400":{"$ref":"#/components/responses/ValidationError"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}}}}